error handling

+++ src/delays/delays.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { Delay, DelayDocument } from './delay.model';
@@ -11,6 +15,10 @@ export class DelaysService {
 
   // Check for an existing delay and update or create a new one
   async setDelay(tripId: string, delay: number): Promise<Delay> {
+    if (delay === 0) {
+      throw new BadRequestException('Delay cannot be set to 0.');
+    }
+
     const existingDelay = await this.delayModel.findOne({ tripId }).exec();
 
     if (existingDelay) {
@@ -29,7 +37,15 @@ export class DelaysService {
     return this.delayModel.findOne({ tripId }).exec();
   }
 
-  async removeDelay(tripId: string): Promise<any> {
-    return this.delayModel.deleteOne({ tripId });
+  async removeDelay(tripId: string): Promise<string> {
+    const existingDelay = await this.delayModel.findOne({ tripId }).exec();
+
+    if (existingDelay) {
+      await this.delayModel.deleteOne({ tripId });
+      return 'Delay removed successfully.';
+    } else {
+      // If no delay found, throw a NotFoundException with a custom message
+      throw new NotFoundException("This train doesn't have a delay.");
+    }
   }
 }
